Separate functions for web and nats startup logic.
