+++ internal/web/about.go
@@ -1,4 +1,4 @@
-package api
+package web
 
 import (
 	"net/url"
